backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
asyncpg==0.30.0
apscheduler==3.10.4
pydantic-settings==2.7.1
